requirements.txt
flask
flask_sqlalchemy
flask_jsonpify
pycryptodome